- Adjusting Inheritance by deleting the PersonID PK and assign driverLic, SSN, SSN to be the PK of the child classes- Customer, Manager, SalesRep respectively - Fixing Adding new Customer, Manager, SlaesRep with same PK / Serive and controller

+++ dms/src/main/java/com/coolfunclub/dms/repository/ManagerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import com.coolfunclub.dms.model.Manager;
 @Repository
-public interface ManagerRepository extends JpaRepository<Manager,Long>{
+public interface ManagerRepository extends JpaRepository<Manager,Integer>{
 
 }
